pyproject.toml
[tool.poetry]
name = "game-api-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "0.111.0"
pydantic = "2.7.4"
uvicorn = "0.30.1"
tortoise-orm = {version = "0.21.3", extras = ["asyncpg"]}
faker = "25.9.1"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
aiogram = "^3.8.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.5"
aerich = "^0.7.2"
telebot = "^0.0.5"

[tool.poetry.dev-dependencies]
flake8 = "6.0.0"
isort = "5.12.0"
black = "23.3.0"


[tool.aerich]
tortoise_orm = "api.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
